wip admindashboard

+++ be/internal/entity/payment.go
@@ -4,7 +4,7 @@ type Payment struct {
 	Id int
 	UserId int
 	User Users `gorm:"foreignKey:UserId;references:Id"`
-	AuctionItemId float64
+	AuctionItemId int
 	Status string
 	// PaymentStatus PaymentStatus `gorm:"foreignKey:StatusId;references:Id"`
 	Amount float64
